api/requirements.txt
flask==2.3.3
redis==5.0.1
psycopg2-binary==2.9.9
